Added A-Packages features
